requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
auto-tune @ file:///root/selfgz159581778/compiler/lib64/auto_tune-0.1.0-py3-none-any.whl
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataflow @ file:///root/selfgz159581778/compiler/lib64/dataflow-0.0.1-py3-none-any.whl
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.0
fsspec==2024.10.0
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
hccl @ file:///root/selfgz159581778/compiler/lib64/hccl-0.1.0-py3-none-any.whl
hccl-parser @ file:///usr/local/Ascend/ascend-toolkit/7.0.0/toolkit/tools/hccl_parser-0.1-py3-none-any.whl
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.7
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msadvisor @ file:///usr/local/Ascend/ascend-toolkit/7.0.0/tools/msadvisor/python/msadvisor-1.0.0-cp37-abi3-linux_aarch64.whl
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
op-compile-tool @ file:///root/selfgz159581778/compiler/lib64/op_compile_tool-0.1.0-py3-none-any.whl
op-gen @ file:///usr/local/Ascend/ascend-toolkit/7.0.0/toolkit/tools/op_gen-0.1-py3-none-any.whl
op-test-frame @ file:///usr/local/Ascend/ascend-toolkit/7.0.0/toolkit/tools/op_test_frame-0.1-py3-none-any.whl
opc-tool @ file:///root/selfgz159581778/compiler/lib64/opc_tool-0.1.0-py3-none-any.whl
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.49
ruff==0.8.0
safehttpx==0.1.1
schedule-search @ file:///root/selfgz159581778/compiler/lib64/schedule_search-0.1.0-py3-none-any.whl
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
supervision==0.25.0
sympy==1.13.1
te @ file:///root/selfgz159581778/compiler/lib64/te-0.4.0-py3-none-any.whl
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.32
ultralytics-thop==2.0.11
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
websockets==12.0
